Minor updates to comments and order of code for clarity (#438)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/438

The file hypothesis_testing.py was previously just for testing means, so, the comment is changed to prepare for the addition of the second test. Also, the hypothesis testing function is placed before the function for constructing the confidence interval, as the first more directly presented the idea of the tests.

Reviewed By: SepehrAkhavan

Differential Revision: D24609581

fbshipit-source-id: cd22408346ae232c0f99637fc6de318c465b722b