Finish initial implementation of JIT compiler (#497)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/497

With this diff the initial attempt at building a JIT compiler is complete; we now handle:

* Calls to random variable functions with ordinary arguments
* Calls to functional functions with any arguments
* Calls to well-known functions with any arguments
* Calls to all other functions with any arguments

As each function is called, we check to see if we have already lifted it; if we need to compile it to the lifted form, we do so, and cache the result.

I've made the executive decision that if we have a call to a function without source code, and the arguments contain graph nodes, we're just going to allow it and hope for the best.

I have not yet implemented support for nested functions or lambdas, and there is no testing yet for models that use classes; I'll handle those in a future diff.

Reviewed By: wtaha

Differential Revision: D25135542

fbshipit-source-id: 3ecfe33675c0836230617fc3fdac7fd9db5de751