Start removing compiler inference hack
Summary: As an initial step along the way to getting inference going in the compiler, I added an "infer" method to the graph builder. I want to redesign that API; step one is renaming the old API so as to not conflict with the new one which will also be named "infer".

Reviewed By: wtaha

Differential Revision: D25288486

fbshipit-source-id: f190fc437a35792de88b0d19f3615735e38eb7b1