Create API for BMG inference from BM models
Summary:
We have now realized Michael's original dream of making Bean Machine Graph "just another inference engine" for a subset of Bean Machine models.   It's a little rough around the edges still, but we actually have an end-to-end solution for the first time with this diff.

To match the pattern established by other inference engines, I've created an object to represent the inference engine; it is a thin wrapper around a BMGraphBuilder which does the actual work:

* jit-compile the functions underlying each query and observation into a lifted form
* execute the lifted form, which populates the graph builder as it runs
* type check the resulting graph for compatibility with BMG
* generate a BMG graph object and run inference on it
* format the resulting sample array into a dictionary of tensors

Still to come:

* Much more testing
* Better error reporting
* Tweaks to BMG and the compiler to support CLARA model
* Improvements to the API such as support for logging, multiple chains, and so on.

Reviewed By: wtaha

Differential Revision: D25291072

fbshipit-source-id: 3783bcef01dfbdca4c2e1d08bbc2453a2abf3c59