Rewrite linear regression test to use BMG inference (#503)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/503

I've rewritten the BMG linear regression test to actually be a genuine end-to-end test that uses BMG inference, both as a random number generator to create a test case, and then to do inference to verify that we get sensible results.

Since we have 200+ samples in this realistic test (100 x, y coordinate pairs) I've removed the test which generates a graph visualization, generates the underlying graph building code, and so on.  We just do the inference as a black box and verify we get sensible results out.

Reviewed By: wtaha

Differential Revision: D25442643

fbshipit-source-id: 0957e1e1aaaa0782f139f5eefcfc400a675f6821