Rename end_to_end_test.py (#504)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/504

In this diff I am just renaming end_to_end_test.py because that name is no longer accurate -- or, maybe, was never accurate. These tests demonstrate that we can translate Bean Machine models in strings into BMG graphs, but they are not true end-to-end scenarios because we are not doing inference.

I'm going to rewrite these tests to use the new entrypoints I've been adding as a result of making the JIT compilation step; I will make those changes in subsequent diffs.

Reviewed By: wtaha

Differential Revision: D25443242

fbshipit-source-id: 3d745a620adf58d0b1502a9648a858862a702ce1