Add debugging helpers for JIT compiler (#505)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/505

When we JIT compile a function, I now stash the original function, the graph builder that asked for the JIT compilation, the AST of the transformed function, and the source code of the transformed function into the transformed function. There is no user-facing feature that requires any of this; it is solely as an aid so that I can easily inspect the operation of the jitter while debugging it.

Reviewed By: wtaha

Differential Revision: D25447511

fbshipit-source-id: 8b359958ad8892cb3516e56b90dcbf82b3e09a65