Fix up graph accumulation tests to use jitter (#507)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/507

The graph accumulation tests now use the jitter functionality on real live Python functions, rather than compiling them as strings, and boy is that more pleasant to read.

I've also cleaned up these tests to standardize on the same test regimen for each:

* declare some random variables that test some functionality in the compiler
* ask for a Graph object to be created on the basis of queries of those random variables
* dump the Graph object as text to verify that we got a valid BMG graph out of the graph accumulator

That more directly tests the "accumulate a graph and transform it to BMG" functionality than also showing how to generate a Python or C++ program that does the same.

Reviewed By: wtaha

Differential Revision: D25444128

fbshipit-source-id: fd0585e11299f456e918e3a5e67c824619d10ad3