Create test case for simplified CLARA model (#509)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/509

The jit compiler can now handle analyzing a basic unvectorized CLARA model; this is the flavor of the model that uses straight probabilities rather than negative real log probabilities.

We give an error when analyzing this model because we in several places add together two probabilities but they sum to a real; BMG does not recognize that a sum of two probabilities may be known to be less than 1.0.

I'll work on solving the type system problem later in this stack of diffs; until then, we'll have the test case check that we are raising an error.

Reviewed By: wtaha

Differential Revision: D25510199

fbshipit-source-id: a2a89636ac3d46f214f81404849ee4152812dc5c