Add TO_PROBABILITY operator to BMG
Summary:
BMG already has a TO_POSITIVE_REAL and TO_REAL nodes, but they both require that their operand already be known to fit into their type; that is, TO_POSITIVE_REAL turns a probability into a positive real, and TO_REAL turns a probability or positive real into a real.

We need a TO_PROBABILITY node that goes the other way: it turns a positive real or real into a probability.  We have need of this because it is common in a model to make an expression which you know and I know must be between 0.0 and 1.0, and needs to be used as a probability, but which the BMG type system will classify as a real or positive real.

Here are some examples; suppose PR is a positive real quantity, P is a probability:

* `0.4 * P + 0.5` -- we know this must fit into a probability, but BMG treats a sum of probabilities as a positive real.
* `1.0 - (0.4 * P + 0.5)` -- again we know that this must fit into a probability, but the parenthesized expression is a positive real according to BMG, and now we have the difference of positive real 1.0 and a positive real, which is a real according to BMG.
* `PR / (1 + PR)`  the quotient of two positive reals is a positive real according to BMG, but we know that this is between 0.0 and 1.0.

We could add operators for every special case we come up with, but that would be a lot of operators and we would undoubtedly miss some. Instead we can simply add a TO_PROBABILITY operator that takes a real or positive real and constrains it to the range of a probability for either (1) cases where we know that we will get a probability, or (2) cases where we don't get a probability but we wish to clamp a real quantity to that range.

Any thoughts and critiques would be very welcome.

Reviewed By: wtaha

Differential Revision: D25547971

fbshipit-source-id: 4bd162a43f97846e89ccabebe7ca5a4e713b2d79