Rename "children" to "inputs"
Summary:
In abstract syntax trees we always describe the inputs to an operation as its "children", but in the exciting world of Bayesian networks we describe inputs as "parents". Unfortunately I learned that after I wrote the graph accumulator.

For performance reasons we are going to need to track both inputs and outputs of each graph node, so now is a good time to make the renaming refactoring that turns "children" into "inputs". I'll add outputs in a later diff.

Reviewed By: wtaha

Differential Revision: D25670130

fbshipit-source-id: a9693979acd786f3dff0b41ef4ff9076a82bb102