Small improvement to item counter wrapper class (#516)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/516

I'm going to need to keep track of the output nodes for each node in the graph builder, not just the input nodes. But it is possible for one node to output to the same node *twice*; if we have 1 + 1, for instance, then the "1" node is an input of the "+" node twice.

The set of outputs will change as the graph is edited and we must maintain the invariant that the output counts are correct. To prepare for this situation I've added item removal from my little item counting wrapper class.

Reviewed By: wtaha

Differential Revision: D25673672

fbshipit-source-id: d57ad6eee6e1edffd6dee637c5312b6c4dd3ab65