Track outputs in all graph builder nodes
Summary:
We now keep track of both inputs and outputs in all nodes; we are not yet consuming the outputs for anything but we will in an upcoming diff.

As noted in the test case, there are some tricky situations we've got to be aware of, such as a single node being an output to two different inputs of the same consuming node; if an edit erases one of those edges then we need to keep track of the remaining edge still.

Reviewed By: wtaha

Differential Revision: D25676400

fbshipit-source-id: fe8de919e398b9da26b596955beac3565855ede5