Refactoring graph type computations
Summary:
To compute the type of a node we might need to know the types of all of its inputs -- but those might need the types of their inputs, and so on.  This means that we may traverse a large fraction of the graph every time we need a graph or inf type, and that happens a lot.

The smart thing to do would be to cache the result after computing it once. However we then run into a problem: the graph might mutate as we are fixing problems in it, and that might change the type of an *output* node; that change might then propagate to its output nodes, and so on.

The solution is to create some invariants:

* every node caches its own type.
* the input nodes of every node have cached their correct type.
* when an input edge of a node is modified to point to a different input node then we recompute the type of the node. If that causes its type to change, then we recompute the types of all its output nodes, and so on.

These invariants guarantee that (1) all reads are fast, and (2) all writes which change a type aggressively propagate the consequences of the change to as few descendent (output) nodes as possible.

But in the code now we recompute the node type on demand. How are we to establish these desirable invariants?

The first step is this relatively straightforward refactoring. Instead of having a virtual property to compute the type, we have a virtual method that explicitly computes the type.

In an upcoming diff we'll make use of this separation to establish these invariants.

Reviewed By: wtaha

Differential Revision: D25677798

fbshipit-source-id: ef1bb94ee6d567012eb86eb862f70131e79b7e4a