Implement type caching for graph nodes (#519)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/519

We now aggressively compute and cache the types of graph nodes. This is for performance reasons. Previously we would recompute the type of a graph node on demand based on its inputs; this could traverse large parts of the graph and be quite expensive, but "what is the type of this node?" is a common operation so it needs to be cheap.

However, caching then presents a difficulty because the graph is mutable; a change in a graph node that changes its type could change the types of its outputs, and so on.  Updating types of outputs transitively is potentially expensive also, but (1) this operation is rare, and (2) usually the mutation does not actually change the type of very many nodes, so the traversal stops quickly.

This fix turns type checking on a typical model from superlinear to linear in the number of mutations, which is much better performance.

Reviewed By: wtaha

Differential Revision: D25680425

fbshipit-source-id: ed08af5168cc9a25c558c857f63de4a7cfcb8fc1