CLARA CGM test for BMG compiler
Summary:
This is a test case based on Kinjal's recent work on a simplified CLARA model; I've rewritten the code slightly to make it more like a tutorial.

This model needs the TO_PROB operator that has not been fully implemented yet, so it does not compile correctly. When we compile correctly I will update the test.

Experimenting with this model led me to discover that we have a superlinear algorithm in the type checker that makes it impossible to analyze graphs with a few hundred nodes in a reasonable time. I believe I know how to fix the performance to put a smaller bound on it; I'll implement that fix later in this stack and then re-enable the full item set in the test.

Reviewed By: wtaha

Differential Revision: D25600369

fbshipit-source-id: a8880c84fbe151a9f31816232e55265371c48af0