CLARA CGM test for BMG compiler (#520)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/520

This is a test case based on Kinjal's recent work on a simplified CLARA model; I've rewritten the code slightly to make it more like a tutorial.

This model needs the TO_PROB operator that has not been fully implemented yet, so it does not compile correctly. When we compile correctly I will update the test.

Experimenting with this model led me to discover that we have a superlinear algorithm in the type checker that makes it impossible to analyze graphs with a few hundred nodes in a reasonable time. I believe I know how to fix the performance to put a smaller bound on it; I'll implement that fix later in this stack and then re-enable the full item set in the test.

Reviewed By: wtaha

Differential Revision: D25600369

fbshipit-source-id: 7fb36149096a89a5f0b6361ce8b30396284601e2