Insert TO_PROBABILITY nodes as required (#521)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/521

As noted in previous diffs in this stack, we occasionally have situations where BMG would analyze an expression as a positive real or real, when we know that it is a probability.  I've added code to the problem fixer which detects this situation and inserts a TO_PROBABILITY node.

This gets both CLARA model test cases working, though (1) inference results are not yet tested in the "complicated" case, and (2) inference results are not as expected in the "simplified" case -- the prevalence estimate seems way off.  I'd expect it to be around 50%, not 2%.  We will investigate why that is in the new year.

I've added a unit test for the problem fixer as well.

Reviewed By: wtaha

Differential Revision: D25687148

fbshipit-source-id: bad58c2ed7a49c1635d397a194ec6717c92b3583