Start re-enabling disabled tests
Summary:
We have a bunch of disabled or partially disabled compiler tests; in order to get a handle on our technical debt at the start of the year, I'm going to do a pass over them and re-enable all the ones that can be easily re-enabled. I'll also remove as many TODO comments as possible.

In this first one: we stopped supporting tensor operations in BMG so I've reworked the first graph accumulator unit test to not use tensors.

Reviewed By: wtaha

Differential Revision: D25772874

fbshipit-source-id: 6fae5a02ae210ccc16deae926a39c2b3e81401ad