Start re-enabling disabled tests (#522)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/522

We have a bunch of disabled or partially disabled compiler tests; in order to get a handle on our technical debt at the start of the year, I'm going to do a pass over them and re-enable all the ones that can be easily re-enabled. I'll also remove as many TODO comments as possible.

In this first one: we stopped supporting tensor operations in BMG so I've reworked the first graph accumulator unit test to not use tensors.

Reviewed By: feynmanliang, wtaha

Differential Revision: D25772874

fbshipit-source-id: 7e2fef1183cb8f22d67d0673154e730a79bf68a3