Remove unnecessary node-to-boolean feature from graph accumulator (#524)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/524

When I was first designing the graph accumulator I thought it might sometimes be necessary to have a node which represented a constant have the same truth value as the constant; this was as part of an early experiment in stochastic control flows.

That went nowhere, but the code was still in the product. I've deleted it and the test case which verified it.

Reviewed By: feynmanliang, wtaha

Differential Revision: D25773689

fbshipit-source-id: 5ec808043a6974c0eb592fa780b17e3d82d5bb93