Remove unnecessary TODO comments from unit test (#526)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/526

The unit tests for the graph accumulator had a lot of TODO comments suggesting ways that they could detect errors when compiling a graph. However, detecting these particular errors would be both difficult and pointless.

Basically the idea here is that the graph accumulator is slightly more lenient when intercepting a call to a method such as `Tensor.add` to generate a graph node than the real `Tensor.add` is; the real `Tensor.add` would produce an error but the compiler allows it.

It would be a pointless waste of time to detect these errors for several reasons.

First, the scenario in question doesn't happen in the first place. If we detect a call to `Tensor.add` and neither of the operands are graph nodes then we just do the call, so if it crashes, it will still crash.

Second, the premise of the scenario is that the user would have to be compiling a model that crashes when you run it as a Bean Machine model. Users shouldn't do that in the first place, but if they do, there is no requirement that we produce the same error! Automatically fixing a mistake and making a semantically equivalent model is not a bad thing.

Reviewed By: feynmanliang

Differential Revision: D25785696

fbshipit-source-id: f62edee726f27bef1d5aae2646d3dca4800998ba