Move cycle detection to its own test, make end-to-end
Summary: We detect in the compiler when a model contains a cycle and therefore cannot be rendered as a DAG; I've moved the tests for that into their own unit test module. Moreover, they now are "end to end style", operating on actual objects declared in the test rather than on a string containing the source code.

Reviewed By: feynmanliang

Differential Revision: D25786562

fbshipit-source-id: 3db85464d947858e51b31c772470b8a6add3126e