Add type checking on inference APIs
Summary:
If you pass in something other than a random variable as a query, or something other than a tensor as an observation, inference eventually crashes with an obscure and unhelpful error message.

Our primary API should be more user-friendly and let the novice know they've done something wrong.

I personally am prone to writing `[flip]` instead of `[flip()]` which gives the unhelpful error that `function` does not have an attribute `function`.

We now interrogate the queries and observations to ensure they are properly typed and give a helpful error if they are not.

Reviewed By: feynmanliang

Differential Revision: D25790657

fbshipit-source-id: e0c39c1e1b38ddacd61fe618d2c42806bda7cd4f