Rework BMA test model for compiler to use public API (#531)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/531

I've changed the BMA model in the compiler test cases to use the public `BMGInference` API rather than passing the source code around as a string.

A nice side effect of this change is doing so apparently now picks a different random seed in the inference engine so we get slighly more accurate results out.

This test case now makes redundant two of the original compiler test cases, so I've removed them. One of those tests had already been disabled because it was crashing the compiler; I isolated the cause of the crash and rewrote the test to exercise just that line. I'll track down the cause of that crash next.

Reviewed By: wtaha

Differential Revision: D25794619

fbshipit-source-id: cf9b1b3497699416787fe035b736495e9a60fa2d