Fix transformation of function calls (#533)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/533

In rewriting the model we wish to turn all the model's function calls into calls to `handle_function`; that way we can intercept calls to functions which require us to accumulate nodes into the graph.

Previous to this diff a significant amount of the transformation was done by the model rewriter itself, but recent changes to the single assignment rewriter make this code both unnecessary and wrong. This diff corrects that error.

The new single assignment rewriter ensures that every call in the rewritten function is of the form: `x = f(*args, **kwargs)` where `args` and `kwargs` are identifiers, or some simple `dict(...)` constructions.  We can ignore the second form entirely; there is never a need to intercept the creation of a simple dictionary; it is only the first form we need to intercept and the program is already in the form in which it is easiest to transform. We now transform from:

    x = f(*args, **kwargs)

to

    x = bmg.handle_func(f, args, kwargs)

Easy peasy.

This fixes a crashing bug where we incorrectly rewrote `dict(**id, **id)`; we now skip that rewrite entirely.

Reviewed By: wtaha

Differential Revision: D25837811

fbshipit-source-id: 0a0954caa509bb98e4b2d087d52241d98e19c11d