Rename TensorNode to ConstantTensorNode (#534)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/534

In order to compile models containing certain multiary operators (such as `logsumexp`), we will need nodes in the graph accumulator which represent a tensor whose elements are other graph nodes.  For example, we might have Bean Machine model code like:

    a = normal(1)
    b = normal(2)
    c = normal(3)
    s = tensor([a, b, c]).logsumexp(dim=0)

where a, b, c are sample nodes in the graph. We do not currently support calling the `tensor()` constructor with graph nodes as arguments but we will need to in order to make this work.

I want the type in the node hierarchy to be named `TensorNode`, but we already use that to represent a *constant* tensor.  (Note: at present we do not actually generate constant BMG matrix node from constant tensors, but we will eventually need to support this scenario. We'll get that working later.)

Therefore in this diff I am just renaming `TensorNode` to `ConstantTensorNode`. In an upcoming diff I will add a new type `TensorNode` to represent a tensor with graph nodes for elements.

Reviewed By: wtaha

Differential Revision: D25842188

fbshipit-source-id: 0a7a2bd76b6874b583fa561b1c81b0e8a5c04ddc