Refactor to make input edge names computed on demand
Summary:
The graph accumulator keeps track of the input and output edges of every node in the graph, and knows the names of all the input edges.

Previously we just stuck that in an attribute of the class because the vast majority of the time, the set of names is fixed and known. However, I am planning on adding an arbitrary-sized stochastic tensor node which could have a large number of input nodes; there is no need to compute the names of those edges until we need them. (To display the graph, for instance, or to produce an error message.)

I've therefore refactored slightly. `edges` is now a wrapper property around an attribute `_edges` which can either be initialized normally in the derived class declarations, or on demand in a virtual method.

Reviewed By: wtaha

Differential Revision: D25851902

fbshipit-source-id: e8a7ea36ad4b5a1d01e197ae283997ae2c45da9a