Start implementing a stochastic tensor node (#540)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/540

In order to make operators like logsumexp work, we'll need to first be able to represent a graph node which is logically a tensor whose elements are also graph nodes.  This diff is a first step in that direction.

Longer term we will need to add support for this sort of node in BMG itself; when we do that we will need to do work to make sure it complies with the type requirements of BMG.  But for now all we need to do is get the bare minimum necessary working for the CLARA model.  Thus, I've left the type analysis marked as TODO, along with a few other work items that are not necessary at this time.

This diff just creates the node class itself. In upcoming diffs we will:

* create the methods in the graph accumulator that add the node to the graph
* map the tensor() constructor to methods that create this node.
* build a node to represent LogSumExp operations
* create methods to add that node to the graph
* map the logsumexp() method
* test the whole thing end to end
* start working on related operators that operate on the values of a tensor

Reviewed By: wtaha

Differential Revision: D25895112

fbshipit-source-id: bde8597910e78aac8320057e445b4f010abc000b