Fix to failing nightly test - T83082230
Summary: This diff fixes a failing nightly test. Basically, an  incidental change to testing code appears to have caused the random sample used in the test to change. The failing test was a result of a standard deviation being somewhat lower than the alpha confidence interval for this value. Such failures can easily occur as low-probability events when a sample changes. A small increase in the number of samples was made, and the test passed when this was done. Increasing the sample size is generally a safe way to respond to this kind of failing test, as it is unlikely to mask a serious underlying problem.

Reviewed By: yucenli, horizon-blue

Differential Revision: D25958740

fbshipit-source-id: a5748f7aa9080b9c13fed51b3057ddeca9723585