Add method to graph builder for creating tensor nodes.
Summary:
We are going to need to represent tensors that contain graph nodes while accumulating the graph. This is the next step in that change; we are simply adding a method to the graph builder that creates a graph node.

Note that as with all graph accumulation operations, adding an existing node simply re-uses that node. The graph is automatically deduplicated.

{F360122181}

Reviewed By: wtaha

Differential Revision: D25930049

fbshipit-source-id: de20494365111e91be1eeef5ca763d0780a5b15b