Remove restriction on query nodes in graph builder (#547)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/547

BMG requires that a query on a graph node be placed exclusively on operator nodes. (Note that sample is an operator.) When creating a type in the graph accumulator to represent queries I therefore added a Python type annotation requiring the operand of a query to be an operator.

However, it is possible for a queried random variable to be a constant or other non-operator value as the graph is accumulated. This should be an error, but it should be an error when the BMG graph is built, not when the graph is accumulated.

I have not yet added the tests or the error reporting for these scenarios; a necessary first step is removing the restriction that a query node only be added to refer to an operator node. This diff is that first step; the remaining steps will happen in upcoming diffs.

Reviewed By: wtaha

Differential Revision: D25982797

fbshipit-source-id: 7f353dcb1b54c083eaf9fc6fdb561340665cf7d7