Compile tensor constructions with graph node elements
Summary:
We now detect and intercept constructions of tensors whose elements are graph nodes and turn them into graph nodes themselves.

We do not (yet) have such a graph node in BMG, but we will need this node as an intermediate stage for analyzing models which contain operations that run on an entire tensor, such as logsumexp. That support will be added in an upcoming diff.

Differential Revision: D25984058

fbshipit-source-id: 07b1d4f5119795fba508445f21cc0ff1b933f5dc