Compile tensor constructions with graph node elements (#548)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/548

We now detect and intercept constructions of tensors whose elements are graph nodes and turn them into graph nodes themselves.

We do not (yet) have such a graph node in BMG, but we will need this node as an intermediate stage for analyzing models which contain operations that run on an entire tensor, such as logsumexp. That support will be added in an upcoming diff.

Reviewed By: wtaha

Differential Revision: D25984058

fbshipit-source-id: bdb3f52b551c4dbc52c21444707dfc5e676d2766