Start implementing logsumexp operator (#549)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/549

I've started implementing the logsumexp operator.

The problem we're going to have with this operator is that in the original python model, the typical usage will be something like:

    y = tensor([norm(1), norm(2), ... ]).logsumexp(dim=1)

That is, we have an explicit tensor constructed from samples and we wish to efficiently and accurately compute the logsumexp.  But in BMG we do not have a "tensor made up of samples" node; we just have a logsumexp node whose inputs are the samples.

We will therefore need to *accumulate* the graph in the first form but *output* it in the second form. We therefore have a somewhat multi-purpose `LogSumExpNode` implementation here; the expectation is that we will construct it in the "stochastic tensor as argument" form and then transform it to eliminate the unsupported tensor node from the BMG graph.

That will all come in upcoming diffs.  For now I'm just creating the node and computing the type requirements of each input.  The type requirements are based on the final form, not on the original form.

Reviewed By: wtaha

Differential Revision: D26007732

fbshipit-source-id: a10c78d362ed1e6b4c3f6b202e2aa0b26855cafb