Support compiling logsumexp operator to BMG
Summary:
In previous diffs we accumulated tensor(...) operations and logsumexp(...) operations into the graph but did not translate them into BMG; before doing so we need to do a transformation to remove the tensor operation.

This transformation is extremely simple; it requires that the input to the logsumexp operation be a **directly constructed** tensor. That is, we now support compiling graphs for models like:

    t = tensor([norm(1), norm(2), 1.25])
    lse = t.logsumexp()

because t is the direct result of a tensor construction. But we do NOT yet support something like

    t = tensor([norm(1), norm(2), 1.25])
    u = tensor([norm(3), norm(4), 1.25])
    v = t + u
    lse = v.logsumexp()

because `v` is the result of a mathematical operation on constructed tensors. We could support this, and we should support it, but that is a future work item. Right now we support this simplified scenario to unblock the CLARA model.

Reviewed By: wtaha

Differential Revision: D26024128

fbshipit-source-id: 9805034e5cd18bfeffdee9ac8284466091b738be