Prevent compiler from crashing when given unsupported code container (#562)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/562

The compiler which produces the "lifted" form of a function was assuming that it would always be given a normal function or method, and was crashing when given a class constructor. This diff causes the compiler to skip code containers that it does not support; we will assume that those containers do not need to be compiled for now, and add support to them later.

Reviewed By: kshah1997

Differential Revision: D26118433

fbshipit-source-id: 58d168543d30fc128aca79b936cf615830f27fee