Prevent compiler from crashing when given unsupported code container
Summary: The compiler which produces the "lifted" form of a function was assuming that it would always be given a normal function or method, and was crashing when given a class constructor. This diff causes the compiler to skip code containers that it does not support; we will assume that those containers do not need to be compiled for now, and add support to them later.

Differential Revision: D26118433 (https://github.com/facebookincubator/beanmachine/commit/6f4a3e8f14ad22a6d7a6dc143fef4b8fa67d8b2a)

fbshipit-source-id: 670af38b1af07caed577397222faf4c0b02babf7