Fix jit compilation of class methods (#568)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/568

In Python a call to a class method has an interesting implementation:

    class C:
      def m(self, x):
        ...
    c = C()
    cm = c.m
    cm(1)

Every time you have `c.m` in a Python program, what happens is Python constructs a new object, here stored in variable `cm`.  When you call `cm(1)`, what it really does is calls:

    cm.__func__(cm.__self__, 1)

where `cm.__func__` is equal to `C.m`.

The JIT compiler was not handling this situation correctly; what we should do is when we encounter a call to `cm(1)` here, we should JIT-compile `C.m` and then call the jitted helper function as `helper(cm.__self__, 1)`.  We were omitting the `cm.__self__` and getting an error.

We now canonicalize calls on a method object to be calls on the original function object and add `cm.__self__` as the first argument.

Reviewed By: kshah1997

Differential Revision: D26154119

fbshipit-source-id: 2bb1e41fd8e27cc30ee7a0be3a1490a4a7882a89