Implement support for alternative "phi" syntax in compiler
Summary:
When compiling a BM model to BMG we detect usages of the "phi" function; that is, the CDF of Normal(0.0, 1.0).  For example we would detect:

    n = Normal(0.0, 1.0)
    x = n.cdf(some_rv())

but we did not detect the alternative syntax:

    n = Normal(0.0, 1.0)
    x = Normal.cdf(n, some_rv())

This diff fixes that oversight by moving the "is this phi?" detection to *after* function calls are canonicalized from the first form to the second.

Reviewed By: wtaha

Differential Revision: D26183153

fbshipit-source-id: e52deead518132bd977326126775c306e470408d