Implement support for alternative "phi" syntax in compiler (#574)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/574

When compiling a BM model to BMG we detect usages of the "phi" function; that is, the CDF of Normal(0.0, 1.0).  For example we would detect:

    n = Normal(0.0, 1.0)
    x = n.cdf(some_rv())

but we did not detect the alternative syntax:

    n = Normal(0.0, 1.0)
    x = Normal.cdf(n, some_rv())

This diff fixes that oversight by moving the "is this phi?" detection to *after* function calls are canonicalized from the first form to the second.

Reviewed By: wtaha

Differential Revision: D26183153

fbshipit-source-id: c62cd5eabb078daa128bb67a2de96a4ad1c033ff