Support for testing sequences of rewrites
Summary:
This diff introduces a new method to support checking of sequences of rewrites (by a specific rewrite set) rather than just one step of rewriting.

Of course, a test case has been added for this new method, and it illustrates how we expect to use this method. There are already many examples in the same test file where this method can be used to shorten the code. But we will use it first for an upcoming (new) set of rewrites.

Differential Revision: D26192159

fbshipit-source-id: 7e292dd1ea8172106335139dc3de5775c769581b