Add support for negative real inputs to expm1 in BMG (#578)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/578

Bean Machine Graph has an EXPM1 operator with the semantics `exp(x) - 1`.  Previous to this diff it required that its input be either real or positive real, and its output is the same type.

We can improve that slightly by noting that if its input is negative real, then its output is also negative real.

Reviewed By: wtaha

Differential Revision: D26186949

fbshipit-source-id: 127948a83d17b429e96bface7f76bfd52c9bb1bb