Add support for expm1 operator to BM->BMG compiler (#579)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/579

(Normally I split adding a new operator up into multiple diffs but we've been down this road so many times now that I feel OK about doing it all at once, even though that makes a rather large diff.)

In this diff I add support to the Beanstalk compiler for the `expm1()` operator, which has the semantics of `exp(x) - 1`.

We support this operator directly in BMG, and so it is straightforward to turn applications of this function in a model into the BMG node.

Note that we do not detect instances of, say `rv().exp() - 1.0` and turn them into `expm1` nodes. We could do so by adding appropriate code to `fix_problems.py`. If `rv()` is of type negative real or positive real, it might be worthwhile to do so because we can get a more precise type analysis in these cases.

Reviewed By: wtaha

Differential Revision: D26187219

fbshipit-source-id: ca876a746a29d5c0e7060f3845197822e7a4cd1b