Test case for waiting terms (#577)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/577

This diff introduces some additional test cases for the upcoming rewrites (assignment left value expressions). Here we are trying a new idea for supportive tests, namely, capturing explicitly some examples of a property that we would like the rewrites to respect: They should only produce terms where the lhs of the assignment is evaluated after the rhs (which is how Python does things: It evaluates the rhs of an assignment before the lhs).

To see how the tests in this diff check this property we also need to recall that all of our rewrites "emit" new named expressions only immediately outside the current expression to see how these test cases support this property. In practical terms the property is checked by making sure that terms that have a lhs that can be reduced will only be rewritten after the rhs has been fully reduced.

Reviewed By: ericlippert

Differential Revision: D26208010

fbshipit-source-id: cb0c804ef73b39b785919b35313d5365e481b141