Refactor problem fixer into many smaller classes (#586)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/586

The code which looks at an accumulated graph and either transforms it into a form that is compatible with the BMG type system, or gives an error if it is not, was getting too complicated for my liking.

I've split it up into five classes, each of which does one thing, and then we just activate each of the five in order.

We might in the future need to have more complex workflows here, where some passes need to be run more than once, or where we need to track whether progress has been made.  It will be easier to do that sort of work if the fixers are already broken up into smaller pieces like this.

Reviewed By: wtaha

Differential Revision: D26208322

fbshipit-source-id: 5320eec2717e8e33e0a91baef4d9eb3f3a6ac94a