Move categorical-based CLARA test to own module
Summary:
The original version of our CLARA model used categorical distributions; I've moved the test for it to its own module.

Note that we do not yet support JIT compilation of models with stochastic control flows of the form `some_rv(some_discrete_rv())`.  I will fix that in an upcoming diff.

Reviewed By: wtaha

Differential Revision: D26211316

fbshipit-source-id: 4d05e40321a97b0ca1b5a0c32cf8ddf5bbcc6b97