Move categorical-based CLARA test to own module (#591)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/591

The original version of our CLARA model used categorical distributions; I've moved the test for it to its own module.

Note that we do not yet support JIT compilation of models with stochastic control flows of the form `some_rv(some_discrete_rv())`.  I will fix that in an upcoming diff.

Reviewed By: wtaha

Differential Revision: D26211316

fbshipit-source-id: b0c96557e093ae918f89b14abe8aedcf7260759d