Fix to MVN scale in NMC proposer (#566)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/566

As we discovered while discussing D25874172 (https://github.com/facebookincubator/beanmachine/commit/d83cea68149b3b2911a4fc18f909aeb05d73805b), the value passed to the multivariate normal's `scale_tril` parameter is not the correct Cholesky factor for the inverse of the Hessian. This replaces that with the correct factor that is obtained using the properties of a permutation matrix (derivation in comment).

Reviewed By: feynmanliang, horizon-blue

Differential Revision: D26162424

fbshipit-source-id: c9394cf5163452ea70ebc4853ad5f6367840776a